requirements.txt
streamlit
openai-whisper
torch
librosa
soundfile
numpy
ffmpeg-python
matplotlib